Temporarily move datetime ITs back to standalone mode
